+++ packages/isaacscript-common/src/functions/level.ts
@@ -2,14 +2,14 @@ import { DoorSlot } from "isaac-typescript-definitions";
 import { game } from "../core/cachedClasses";
 import { getEnumValues } from "./enums";
 import { isDoorSlotValidAtGridIndexForRedRoom } from "./levelGrid";
-import { getNumRooms, getRoomsInGrid } from "./rooms";
+import { getNumRooms, getRoomsInsideGrid } from "./rooms";
 
 export function fillLevelWithRedRooms(): void {
   const level = game.GetLevel();
 
   let numRoomsInGrid: int;
   do {
-    const roomsInGrid = getRoomsInGrid();
+    const roomsInGrid = getRoomsInsideGrid();
     numRoomsInGrid = roomsInGrid.length;
 
     for (const roomDescriptor of roomsInGrid) {