+++ packages/docs/docs/main/change-log.md
@@ -26,6 +26,9 @@ This page lists the changes to the IsaacScript framework.
   - `getRoomDataForTypeVariant`
   - `setBackdrop`
   - `reloadRoom`
+- Renamed the following helper functions:
+  - `isRoomInsideMap` --> `isRoomInsideGrid`
+  - `getRoomsInGrid` --> `getRoomsInsideGrid`
 - Changed the following helper functions:
   - `spawnCustomGridEntity` now supports the `breakable` argument, which will trigger `POST_GRID_ENTITY_CUSTOM_BROKEN` if it is hit by any type of explosion.
 - Added the following constants: